requirements.txt
flask
flask-CORS
pillow
opencv-python
groq
python-dotenv
werkzeug
gunicorn